% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dwh_connect.R
\name{dwh_connect}
\alias{dwh_connect}
\title{Connect to the Data Warehouse}
\usage{
dwh_connect(x = Sys.getenv("TRUST_DWH"))
}
\arguments{
\item{x}{A character string specifying the name or address of the SQL Server.
Defaults to the value of the environment variable \code{"TRUST_DWH"}.}
}
\value{
A database connection object of class \code{DBIConnection}, which can be used
with \code{DBI} and \code{dbplyr} functions to query the data warehouse.
}
\description{
This function establishes an ODBC connection to the specified SQL Server data warehouse.
}
\examples{
\dontrun{
con <- dwh_connect()
dbListTables(con)
}

}
