% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prop_n.R
\name{prop_n}
\alias{prop_n}
\title{Proportion and Count of a Specific Value}
\usage{
prop_n(variable, value_of_interest, rounding_value)
}
\arguments{
\item{variable}{A vector (numeric, character, or factor).}

\item{value_of_interest}{The value to count.}

\item{rounding_value}{Number of decimal places to round the percentage to.}
}
\value{
A character string formatted as "x\% (n = y)".
}
\description{
Calculates the percentage and count of a specific value in a vector and returns it as a formatted string.
}
