% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prop_n_greater.R
\name{prop_n_greater}
\alias{prop_n_greater}
\title{Proportion and Count of Values Greater Than or Equal to a Threshold}
\usage{
prop_n_greater(variable, value_of_interest, rounding_value)
}
\arguments{
\item{variable}{A numeric vector.}

\item{value_of_interest}{A numeric threshold.}

\item{rounding_value}{Number of decimal places to round the percentage to.}
}
\value{
A character string formatted as "x\% (n = y)".
}
\description{
Calculates the percentage and count of values greater than or equal to a specified threshold.
}
