% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/with_no_lock.R
\name{with_no_lock}
\alias{with_no_lock}
\title{Append NOLOCK hint to SQL table reference}
\usage{
with_no_lock(table_name)
}
\arguments{
\item{table_name}{A character string representing the full name of the table
(e.g., \code{"schema.database.table"}).}
}
\value{
A character string of the form \verb{schema.database.table WITH (NOLOCK)}.
}
\description{
This helper function appends the SQL Server \code{WITH (NOLOCK)} hint to a table
reference string, which can be used in \code{dbplyr::tbl()} calls to allow reading
data without locking.
}
\examples{
with_no_lock("BH_DATAWAREHOUSE.dbo.PI_CDE_PERSON_PATIENT")
# Returns: "BH_DATAWAREHOUSE.dbo.PI_CDE_PERSON_PATIENT WITH (NOLOCK)"

}
